+++ libraries/rush-lib/src/api/LastInstallFlag.ts
@@ -28,8 +28,12 @@ export interface ILockfileValidityCheckOptions {
  * @internal
  */
 export class LastInstallFlag {
-  private _path: string;
   private _state: JsonObject;
+
+  /**
+   * Returns the full path to the flag file
+   */
+  public readonly path: string;
 
   /**
    * Creates a new LastInstall flag
@@ -37,7 +41,7 @@ export class LastInstallFlag {
    * @param state - optional, the state that should be managed or compared
    */
   public constructor(folderPath: string, state: JsonObject = {}) {
-    this._path = path.join(folderPath, this.flagName);
+    this.path = path.join(folderPath, this.flagName);
     this._state = state;
   }
 
@@ -71,7 +75,7 @@ export class LastInstallFlag {
   ): boolean {
     let oldState: JsonObject;
     try {
-      oldState = JsonFile.load(this._path);
+      oldState = JsonFile.load(this.path);
     } catch (err) {
       return false;
     }
@@ -123,7 +127,7 @@ export class LastInstallFlag {
    * Writes the flag file to disk with the current state
    */
   public create(): void {
-    JsonFile.save(this._state, this._path, {
+    JsonFile.save(this._state, this.path, {
       ensureFolderExists: true
     });
   }
@@ -132,14 +136,7 @@ export class LastInstallFlag {
    * Removes the flag file
    */
   public clear(): void {
-    FileSystem.deleteFile(this._path);
-  }
-
-  /**
-   * Returns the full path to the flag file
-   */
-  public get path(): string {
-    return this._path;
+    FileSystem.deleteFile(this.path);
   }
 
   /**