+++ libraries/rush-lib/src/logic/PurgeManager.ts
@@ -15,8 +15,9 @@ import { RushGlobalFolder } from '../api/RushGlobalFolder';
 export class PurgeManager {
   private _rushConfiguration: RushConfiguration;
   private _rushGlobalFolder: RushGlobalFolder;
-  private _commonTempFolderRecycler: AsyncRecycler;
   private _rushUserFolderRecycler: AsyncRecycler;
+
+  public readonly commonTempFolderRecycler: AsyncRecycler;
 
   public constructor(rushConfiguration: RushConfiguration, rushGlobalFolder: RushGlobalFolder) {
     this._rushConfiguration = rushConfiguration;
@@ -26,7 +27,7 @@ export class PurgeManager {
       this._rushConfiguration.commonTempFolder,
       RushConstants.rushRecyclerFolderName
     );
-    this._commonTempFolderRecycler = new AsyncRecycler(commonAsyncRecyclerPath);
+    this.commonTempFolderRecycler = new AsyncRecycler(commonAsyncRecyclerPath);
 
     const rushUserAsyncRecyclerPath: string = path.join(
       this._rushGlobalFolder.path,
@@ -40,12 +41,8 @@ export class PurgeManager {
    * the PurgeManager instance is disposed.
    */
   public deleteAll(): void {
-    this._commonTempFolderRecycler.deleteAll();
+    this.commonTempFolderRecycler.deleteAll();
     this._rushUserFolderRecycler.deleteAll();
-  }
-
-  public get commonTempFolderRecycler(): AsyncRecycler {
-    return this._commonTempFolderRecycler;
   }
 
   /**
@@ -55,7 +52,7 @@ export class PurgeManager {
     // Delete everything under common\temp except for the recycler folder itself
     console.log('Purging ' + this._rushConfiguration.commonTempFolder);
 
-    this._commonTempFolderRecycler.moveAllItemsInFolder(
+    this.commonTempFolderRecycler.moveAllItemsInFolder(
       this._rushConfiguration.commonTempFolder,
       this._getMembersToExclude(this._rushConfiguration.commonTempFolder, true)
     );