+++ apps/rundown/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/rundown
 
-This log was last generated on Tue, 29 Nov 2022 01:16:49 GMT and should not be manually modified.
+This log was last generated on Fri, 09 Dec 2022 16:18:28 GMT and should not be manually modified.
+
+## 1.0.233
+Fri, 09 Dec 2022 16:18:28 GMT
+
+_Version update only_
 
 ## 1.0.232
 Tue, 29 Nov 2022 01:16:49 GMT