+++ webpack/loader-load-themed-styles/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @microsoft/loader-load-themed-styles
 
-This log was last generated on Fri, 02 Dec 2022 01:15:42 GMT and should not be manually modified.
+This log was last generated on Fri, 09 Dec 2022 16:18:28 GMT and should not be manually modified.
+
+## 2.0.16
+Fri, 09 Dec 2022 16:18:28 GMT
+
+_Version update only_
 
 ## 2.0.15
 Fri, 02 Dec 2022 01:15:42 GMT