+++ webpack/set-webpack-public-path-plugin/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/set-webpack-public-path-plugin
 
-This log was last generated on Fri, 02 Dec 2022 01:15:42 GMT and should not be manually modified.
+This log was last generated on Fri, 09 Dec 2022 16:18:28 GMT and should not be manually modified.
+
+## 3.3.82
+Fri, 09 Dec 2022 16:18:28 GMT
+
+_Version update only_
 
 ## 3.3.81
 Fri, 02 Dec 2022 01:15:42 GMT